updated description
There was an error in my initial description about analog values and signal strength, so I fixed it; and, because it is counter-intuitive, I highlighted it.

I also have a concern about not using curly braces with the if statement in these examples.

For  a beginner, this might be a source of confusion. When the sketch is run as it is, everything works fine. But, if one were to copy/paste into a larger program without the braces in the if statement the program would crash and the person might not understand why, since it works fine as a standalone sketch.
Just my 2¢