biTToe KY-036 final draft
It took me a bit or research to figure out how to make this sensor do what I thought it should do..

The built-in LED was a pain to monitor so I wired an external LED to pin 13, figuring it could do double duty with no additional code, but I don't have the Fritzing app to draw the schematic. I did include a photo and a video, however.

To toggle the  LED on and off with each touch, I had to use a flag. I couldn't figure out how to make the flag a local variable without re-setting the flag with each iteration of main(), or adding additional variables, and superfluous code which would only waste time and space.

Cheers
~ƒrank